Updating Cron to match * * * * * tab syntax * bug fixing

+++ system/model/plugin.php
@@ -25,8 +25,9 @@ class System_Model_Plugin extends Model
 		$plugin_entries = $this->queryRows("SELECT * FROM " . DB_PREFIX . "plugin_registry WHERE `name` = '" . $this->db->escape($name) . "'");
 
 		foreach ($plugin_entries as $entry) {
-			if (is_file($entry['live_file'])) {
-				$this->message->add("notify", "Removed plugin file $entry[live_file].");
+			//Only Remove symlinked files (in case someone already deleted this file and replaced it)
+			if (is_file($entry['live_file']) && _is_link($entry['live_file'])) {
+				$this->message->add("notify", _("Removed plugin file $entry[live_file]."));
 				chmod($entry['live_file'], 0777);
 				unlink($entry['live_file']);
 			}
